Remove Gearman dependency
The Gearman Jenkins plugin turned out to be troublesome and
the cron jobs used to trigger the Jenkins jobs through the Gearman
server used to get stucked and the solution was to reset the
connection to the Gearman server.

We decided to use the jenkins-cli.jar to directly call the
Jenkins jobs from the bash. The 'Mesos/utils/jenkins-executor-common.sh'
contains minimal business logic to load-balance job runs across
multiple Jenkins servers similarly configured, just like how
the Gearman server is doing.