Change the logic to update 'latest' symlink
The repository owner is now part of the full location
where the logs and binaries are stored for a build.

Thus, we have an extra 'latest' symlink pointing to the
last successful run for a particular repository owner.

This useful when running the mesos-nightly-build job
against forks of Mesos. For example: https://github.com/mesosphere/mesos