Add Python 3 support for the python scripts
This patch adds a new file called "python_compatibility_utils.py"
which contains the imports and methods that work on both
Python 2 and 3. It exposes the same name for functions and modules,
making it easy to use in other parts of the code.

It also adds brackets to the message from print function as that
is mandatory in Python 3.

Adds __init__.py file to be able to import from current directory
without adding the current directory to system path.