Fix test_windows_agent_dcos_dns
Previously the retry logic to resolve DNS records was implemented
in the PowerShell code executed over WinRM.

Unfortunately, this caused problems and the WinRM call never finishes
if something is wrong and the retry PowerShell logic is involved.

This is fixed by simplifying the PowerShell remote command and
implementing the retry logic locally.