Fix Linux pre-provision scripts
Remove the old code to call the bootstrap via postinstall.sh
since it is hard to decide when the ZooKeeper contains the
cluster-id information.

As a quick workaround, the dcos-metrics bootstrap call is added
as an ExecStartPre  to the dcos-telegraf systemd service file.
This will create the required /var/lib/dcos/cluster-id file
when it's available in ZooKeeper, and thus have a healthy
dcos-telegraf service.